+++ Modern JavaScript From The Beginning/Section 7: Asynchronous JavaScript, Ajax & Fetch API/Custom HTTP Library/lib.js
@@ -19,8 +19,52 @@ HTTPLib.prototype.get = function (url, callback) {
     this.http.send();
 };
 // HTTP POST
-HTTPLib.prototype.post = function () {};
+HTTPLib.prototype.post = function (url, data, callback) {
+    this.http.open('POST', url, true);
+    this.http.setRequestHeader('Content-Type', 'application/json');
+
+    let self = this;
+
+    this.http.onload = function () {
+        if (self.http.status === 201) {
+            callback(null, self.http.responseText);
+        } else {
+            callback(`Error: ${self.http.status}`);
+        }
+    };
+
+    this.http.send(JSON.stringify(data));
+};
 // HTTP PUT
-HTTPLib.prototype.put = function () {};
+HTTPLib.prototype.put = function (url, data, callback) {
+    this.http.open('PUT', url, true);
+    this.http.setRequestHeader('Content-Type', 'application/json');
+
+    let self = this;
+
+    this.http.onload = function () {
+        if (self.http.status === 200) {
+            callback(null, self.http.responseText);
+        } else {
+            callback(`Error: ${self.http.status}`);
+        }
+    };
+
+    this.http.send(JSON.stringify(data));
+};
 // HTTP DELETE
-HTTPLib.prototype.delete = function () {};
+HTTPLib.prototype.delete = function (url, callback) {
+    this.http.open('DELETE', url, true);
+
+    let self = this;
+
+    this.http.onload = function () {
+        if (self.http.status === 200) {
+            callback(null, 'Successfully deleted');
+        } else {
+            callback(`Error: ${self.http.status}`);
+        }
+    };
+
+    this.http.send();
+};