+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 7: Asynchronous JavaScript, Ajax & Fetch API
     - Callback Functions
     - Custom HTTP Library (Ajax With Callbacks) - Part 1
+    - Custom HTTP Library (Ajax With Callbacks) - Part 2
 - An example with a callback function was added.
 - A get method was added.
+- Post, put and delete methods were added.
 
 ## September 8, 2020
 - Modern JavaScript From The Beginning (Udemy)